requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.7.0
tqdm>=4.65.0
pyiqa>=0.1.7
lietorch
cvxpy
submitit
structlog
scipy
scikit-image
